requirement.txt
asgiref==3.2.5
Django==3.0.4
django-appconf==1.0.4
django-countries==6.1.3
django-js-asset==1.2.2
django-jsonsuit==0.4.4
django-mptt==0.11.0
django-mptt-urls==2.0.4
django-prettyjson==0.4.1
future==0.18.2
Pillow==7.1.1
psycopg2==2.8.5
pytils==0.3
pytz==2019.3
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
standardjson==0.3.1
